finalizes fix for longitude latitude storage precision

+++ server/db/migrations/20170717181722_add_longitude_latitude_to_events.js
@@ -1,8 +1,8 @@
 
 exports.up = function(knex, Promise) {
   return Promise.all([knex.schema.table('events', function(table){
-    table.float('longitude', 14, 10);
-    table.float('latitude', 14, 10);
+    table.decimal('longitude', 14, 10);
+    table.decimal('latitude', 14, 10);
   })
   ]);
 };
